feat: move translate and open ai service to application generic
