+++ apps/api/package.json
@@ -73,7 +73,6 @@
     "nanoid": "^3.1.20",
     "nest-raven": "^9.2.0",
     "newrelic": "9.10.2",
-    "openai": "^3.2.1",
     "passport": "0.6.0",
     "passport-github2": "^0.1.12",
     "passport-jwt": "^4.0.0",
