+++ apps/api/src/app/messages/usecases/index.ts
@@ -1,5 +1,4 @@
-import { TranslateMessage } from './translate-message/translate-message.usecase';
 import { GetMessages } from './get-messages';
 import { RemoveMessage } from './remove-message';
 
-export const USE_CASES = [RemoveMessage, GetMessages, TranslateMessage];
+export const USE_CASES = [RemoveMessage, GetMessages];
