--- apps/api/src/app/messages/usecases/translate-message/translate-message.command.ts
@@ -1,10 +1,0 @@
-import { IsString, IsIn } from 'class-validator';
-import { EnvironmentCommand } from '../../../shared/commands/project.command';
-
-export class TranslateMessageCommand extends EnvironmentCommand {
-  @IsString()
-  messageContent: string;
-
-  @IsIn(['French', 'Japanese', 'Spanish', 'es-ES'])
-  language: 'French' | 'Japanese' | 'Spanish' | 'es-ES';
-}
