+++ packages/application-generic/package.json
@@ -100,6 +100,7 @@
     "lodash": "^4.17.15",
     "nestjs-pino": "^3.1.2",
     "node-fetch": "^3.2.10",
+    "openai": "^3.2.1",
     "pino-http": "^8.3.3",
     "pino-pretty": "^9.4.0",
     "redlock": "4.2.0",
