+++ packages/application-generic/src/services/index.ts
@@ -5,6 +5,7 @@ export { WsQueueService } from './ws-queue.service';
 export { TriggerQueueService } from './trigger-queue.service';
 export { VerifyPayloadService } from './verify-payload.service';
 export { EventsDistributedLockService } from './events-distributed-lock.service';
+export { OpenAiService } from './openai/openai.service';
 export * from './cache';
 export * from './calculate-delay';
 export * from './performance';
