+++ packages/application-generic/src/usecases/index.ts
@@ -18,3 +18,4 @@ export * from './store-subscriber-jobs';
 export * from './trigger-event';
 export * from './create-notification-jobs';
 export * from './process-subscriber';
+export * from './translate-message';
