+++ packages/application-generic/src/usecases/translate-message/index.ts
@@ -1,0 +1,2 @@
+export * from './translate-message.command';
+export * from './translate-message.usecase';
