+++ packages/application-generic/src/usecases/translate-message/translate-message.command.ts
@@ -1,0 +1,10 @@
+import { IsString, IsIn } from 'class-validator';
+import { EnvironmentCommand } from '../../commands';
+
+export class TranslateMessageCommand extends EnvironmentCommand {
+  @IsString()
+  messageContent: string;
+
+  @IsString()
+  language: string;
+}
