+++ packages/application-generic/src/usecases/translate-message/translate-message.usecase.ts
@@ -1,5 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { OpenAiService } from '../../../shared/services/openai/openai.service';
+import { OpenAiService } from '../../services';
+import { isLanguageSupported } from '../../utils/translate-message';
+
 import { TranslateMessageCommand } from './translate-message.command';
 
 @Injectable()
@@ -7,6 +9,9 @@ export class TranslateMessage {
   constructor(private openAiService: OpenAiService) {}
 
   async execute(command: TranslateMessageCommand) {
+    if (isLanguageSupported(command.language)) {
+    }
+
     const prompt = `Translate the following message into ${command.language}:\n\n"${command.messageContent}"`;
 
     const result = await this.openAiService.translateMessage(prompt);
